+++ Controller/EventsController.php
@@ -21,6 +21,7 @@ use Symfony\Component\HttpFoundation\Session\Session;
 use Symfony\Component\HttpKernel\Exception\HttpException;
 use Cocur\Slugify\Slugify;
 use c975L\EventsBundle\Entity\Event;
+use c975L\EventsBundle\Service\EventService;
 use c975L\EventsBundle\Form\EventType;
 
 class EventsController extends Controller
@@ -55,14 +56,19 @@ class EventsController extends Controller
                 10
             );
 
+            //Defines toolbar
+            $tools  = $this->renderView('@c975LEvents/tools.html.twig', array(
+                'type' => 'dashboard',
+            ));
+            $toolbar = $this->forward('c975L\ToolbarBundle\Controller\ToolbarController::displayAction', array(
+                'tools'  => $tools,
+                'dashboard'  => 'events',
+            ))->getContent();
+
             //Returns the dashboard
             return $this->render('@c975LEvents/pages/dashboard.html.twig', array(
                 'events' => $pagination,
-                'toolbar' => $this->renderView('@c975LEvents/toolbar.html.twig', array(
-                    'type' => 'dashboard',
-                    'dashboardRoute' => $this->getParameter('c975_l_events.dashboardRoute'),
-                    'signoutRoute' => $this->getParameter('c975_l_events.signoutRoute'),
-                )),
+                'toolbar' => $toolbar,
             ));
         }
 
@@ -82,17 +88,20 @@ class EventsController extends Controller
      */
     public function displayAction($id)
     {
-        //Gets the user
-        $user = $this->getUser();
+        //Gets the Service
+        $eventsService = $this->get(\c975L\EventsBundle\Service\EventsService::class);
 
         //Gets the event
-        $event = $this->loadEvent($id);
-        $this->setPicture($event);
+        $event = $eventsService->load($id);
+        $eventsService->setPicture($event);
 
         //Deleted event
         if ($event->getSuppressed() === true) {
             throw new HttpException(410);
         }
+
+        //Gets the user
+        $user = $this->getUser();
 
         //Defines toolbar
         $toolbar = '';
@@ -107,7 +116,7 @@ class EventsController extends Controller
             ))->getContent();
         }
 
-        return $this->render('@c975LEvents/pages/eventDisplay.html.twig', array(
+        return $this->render('@c975LEvents/pages/display.html.twig', array(
             'toolbar' => $toolbar,
             'event' => $event,
         ));
@@ -135,8 +144,11 @@ class EventsController extends Controller
                 //Gets the manager
                 $em = $this->getDoctrine()->getManager();
 
+                //Gets the Service
+                $eventsService = $this->get(\c975L\EventsBundle\Service\EventsService::class);
+
                 //Adjust slug in case of not accepted signs
-                $event->setSlug($this->slugify($form->getData()->getSlug()));
+                $event->setSlug($eventsService->slugify($form->getData()->getSlug()));
 
                 //Persists data in DB
                 $em->persist($event);
@@ -144,7 +156,7 @@ class EventsController extends Controller
 
                 //Resizes and renames the picture
                 if ($form->getData()->getPicture() !== null) {
-                    $this->resizeImage($form->getData()->getPicture(), $event->getSlug() . '-' . $event->getId());
+                    $eventsService->resizeImage($form->getData()->getPicture(), $event->getSlug() . '-' . $event->getId());
                 }
 
                 //Redirects to the event
@@ -154,14 +166,19 @@ class EventsController extends Controller
                 ));
             }
 
+            //Defines toolbar
+            $tools  = $this->renderView('@c975LEvents/tools.html.twig', array(
+                'type' => 'new',
+            ));
+            $toolbar = $this->forward('c975L\ToolbarBundle\Controller\ToolbarController::displayAction', array(
+                'tools'  => $tools,
+                'dashboard'  => 'events',
+            ))->getContent();
+
             //Returns the form to edit content
             return $this->render('@c975LEvents/forms/eventNew.html.twig', array(
                 'form' => $form->createView(),
-                'toolbar' => $this->renderView('@c975LEvents/toolbar.html.twig', array(
-                    'type' => 'new',
-                    'dashboardRoute' => $this->getParameter('c975_l_events.dashboardRoute'),
-                    'signoutRoute' => $this->getParameter('c975_l_events.signoutRoute'),
-                )),
+                'toolbar' => $toolbar,
                 'tinymceApiKey' => $this->container->hasParameter('tinymceApiKey') ? $this->getParameter('tinymceApiKey') : null,
                 'tinymceLanguage' => $this->getParameter('c975_l_events.tinymceLanguage'),
             ));
@@ -188,12 +205,15 @@ class EventsController extends Controller
 
         //Defines the form
         if ($user !== null && $this->get('security.authorization_checker')->isGranted($this->getParameter('c975_l_events.roleNeeded'))) {
+            //Gets the Service
+            $eventsService = $this->get(\c975L\EventsBundle\Service\EventsService::class);
+
             //Gets the event
-            $event = $this->loadEvent($id);
+            $event = $eventsService->load($id);
             $originalSlug = $event->getSlug();
 
             //Gets the existing picture
-            $this->setPicture($event);
+            $eventsService->setPicture($event);
             $eventPicture = $event->getPicture();
             if ($eventPicture !== null) {
                 $event->setPicture(new File($eventPicture));
@@ -211,7 +231,7 @@ class EventsController extends Controller
                 $fs = new Filesystem();
 
                 //Adjust slug in case of not accepted signs
-                $event->setSlug($this->slugify($form->getData()->getSlug()));
+                $event->setSlug($eventsService->slugify($form->getData()->getSlug()));
 
                 //Renames picture file if slug has changed
                 $folderPath = $this->getParameter('kernel.root_dir') . '/../web/images/' . $this->getParameter('c975_l_events.folderPictures') . '/';
@@ -222,7 +242,7 @@ class EventsController extends Controller
 
                 //Resizes and renames the picture (that will erase existing one)
                 if ($form->getData()->getPicture() !== null) {
-                    $this->resizeImage($form->getData()->getPicture(), $event->getSlug() . '-' . $event->getId());
+                    $eventsService->resizeImage($form->getData()->getPicture(), $event->getSlug() . '-' . $event->getId());
                 }
 
                 //Persists data in DB
@@ -236,18 +256,23 @@ class EventsController extends Controller
                 ));
             }
 
+            //Defines toolbar
+            $tools  = $this->renderView('@c975LEvents/tools.html.twig', array(
+                'type' => 'edit',
+                'event' => $event,
+            ));
+            $toolbar = $this->forward('c975L\ToolbarBundle\Controller\ToolbarController::displayAction', array(
+                'tools'  => $tools,
+                'dashboard'  => 'events',
+            ))->getContent();
+
             //Returns the form to edit content
             return $this->render('@c975LEvents/forms/eventEdit.html.twig', array(
                 'form' => $form->createView(),
                 'event' => $event,
                 'eventPicture' => $eventPicture,
                 'eventTitle' => $event->getTitle(),
-                'toolbar' => $this->renderView('@c975LEvents/toolbar.html.twig', array(
-                    'type' => 'edit',
-                    'event' => $event,
-                    'dashboardRoute' => $this->getParameter('c975_l_events.dashboardRoute'),
-                    'signoutRoute' => $this->getParameter('c975_l_events.signoutRoute'),
-                )),
+                'toolbar' => $toolbar,
                 'tinymceApiKey' => $this->container->hasParameter('tinymceApiKey') ? $this->getParameter('tinymceApiKey') : null,
                 'tinymceLanguage' => $this->getParameter('c975_l_events.tinymceLanguage'),
             ));
@@ -274,10 +299,13 @@ class EventsController extends Controller
 
         //Defines the form
         if ($user !== null && $this->get('security.authorization_checker')->isGranted($this->getParameter('c975_l_events.roleNeeded'))) {
+            //Gets the Service
+            $eventsService = $this->get(\c975L\EventsBundle\Service\EventsService::class);
+
             //Gets the event
-            $event = $this->loadEvent($id);
+            $event = $eventsService->load($id);
             $event->setAction('delete');
-            $this->setPicture($event);
+            $eventsService->setPicture($event);
 
             //Defines form
             $form = $this->createForm(EventType::class, $event);
@@ -307,17 +335,22 @@ class EventsController extends Controller
                 return $this->redirectToRoute('events_dashboard');
             }
 
+            //Defines toolbar
+            $tools  = $this->renderView('@c975LEvents/tools.html.twig', array(
+                'type' => 'delete',
+                'event' => $event,
+            ));
+            $toolbar = $this->forward('c975L\ToolbarBundle\Controller\ToolbarController::displayAction', array(
+                'tools'  => $tools,
+                'dashboard'  => 'events',
+            ))->getContent();
+
             //Returns the form to edit content
             return $this->render('@c975LEvents/forms/eventDelete.html.twig', array(
                 'form' => $form->createView(),
                 'eventTitle' => $event->getTitle(),
                 'event' => $event,
-                'toolbar' => $this->renderView('@c975LEvents/toolbar.html.twig', array(
-                    'type' => 'delete',
-                    'event' => $event,
-                    'dashboardRoute' => $this->getParameter('c975_l_events.dashboardRoute'),
-                    'signoutRoute' => $this->getParameter('c975_l_events.signoutRoute'),
-                )),
+                'toolbar' => $toolbar,
             ));
         }
 
@@ -337,8 +370,11 @@ class EventsController extends Controller
      */
     public function icalAction($id)
     {
+        //Gets the Service
+        $eventsService = $this->get(\c975L\EventsBundle\Service\EventsService::class);
+
         //Gets the event
-        $event = $this->loadEvent($id);
+        $event = $eventsService->load($id);
 
         //Returns the ical
         return new Response(
@@ -369,9 +405,12 @@ class EventsController extends Controller
         //Loads from DB
         $events = $repository->findForCarousel($number);
 
+        //Gets the Service
+        $eventsService = $this->get(\c975L\EventsBundle\Service\EventsService::class);
+
         //Assigns picture
         foreach ($events as $event) {
-            $this->setPicture($event);
+            $eventsService->setPicture($event);
         }
 
         //Returns the carousel
@@ -418,7 +457,10 @@ class EventsController extends Controller
      */
     public function slugAction($text)
     {
-        return $this->json(array('a' => $this->slugify($text)));
+        //Gets the Service
+        $eventsService = $this->get(\c975L\EventsBundle\Service\EventsService::class);
+
+        return $this->json(array('a' => $eventsService->slugify($text)));
     }
 
 //HELP
@@ -434,147 +476,22 @@ class EventsController extends Controller
 
         //Returns the dashboard content
         if ($user !== null && $this->get('security.authorization_checker')->isGranted($this->getParameter('c975_l_events.roleNeeded'))) {
+            //Defines toolbar
+            $tools  = $this->renderView('@c975LEvents/tools.html.twig', array(
+                'type' => 'help',
+            ));
+            $toolbar = $this->forward('c975L\ToolbarBundle\Controller\ToolbarController::displayAction', array(
+                'tools'  => $tools,
+                'dashboard'  => 'events',
+            ))->getContent();
+
             //Returns the help
             return $this->render('@c975LEvents/pages/help.html.twig', array(
-                'toolbar' => $this->renderView('@c975LEvents/toolbar.html.twig', array(
-                    'type' => 'help',
-                    'dashboardRoute' => $this->getParameter('c975_l_events.dashboardRoute'),
-                    'signoutRoute' => $this->getParameter('c975_l_events.signoutRoute'),
-                )),
+                'toolbar' => $toolbar,
             ));
         }
 
         //Access is denied
         throw $this->createAccessDeniedException();
-    }
-
-
-//FUNCTIONS
-    //Defines the picture related to Event
-    public function setPicture($event)
-    {
-        //Gets the FileSystem
-        $fs = new Filesystem();
-
-        $folderPath = $this->getParameter('kernel.root_dir') . '/../web/images/' . $this->getParameter('c975_l_events.folderPictures') . '/';
-        $picture = $folderPath . $event->getSlug() . '-' . $event->getId() . '.jpg';
-        if ($fs->exists($picture)) {
-            $event->setPicture('images/' . $this->getParameter('c975_l_events.folderPictures') . '/' . $event->getSlug() . '-' . $event->getId() . '.jpg');
-        }
-    }
-
-    //Loads the event
-    public function loadEvent($id)
-    {
-        //Gets the manager
-        $em = $this->getDoctrine()->getManager();
-
-        //Gets repository
-        $repository = $em->getRepository('c975LEventsBundle:Event');
-
-        //Loads from DB
-        $event = $repository->findOneById($id);
-
-        //Not existing event
-        if (!$event instanceof Event) {
-            throw $this->createNotFoundException();
-        }
-
-        return $event;
-    }
-
-    //Resizes the picture
-    public function resizeImage($file, $finalFileName)
-    {
-        //Defines data
-        $extension = is_object($file) ? strtolower($file->guessExtension()) : substr($file, strrpos($file, '.') + 1, 3);
-        $finalHeight = 400;
-        $format = 'jpg';
-
-        //Rotates (if needed) and resizes
-        if (in_array($extension, array('jpeg', 'jpg', 'png')) === true) {
-            $fileData = getimagesize($file);
-            //Also used to reduces poster issued from video
-            $filename = is_object($file) ? $file->getRealPath() : $file;
-            //Use of of @ avoids errors of type IFD bad offset...
-            $exifData = @exif_read_data($filename, 0, true);
-
-            //Creates the final picture
-            if (is_array($fileData)) {
-                //Defines data
-                $compressionJpg = 75;
-                $width = $fileData[0];
-                $height = $fileData[1];
-
-                //Resizes image
-                $newHeight = $finalHeight;
-                $newWidth = (int) round(($width * $newHeight) / $height);
-                $degree = 0;
-
-                //JPEG format
-                if ($fileData[2] == 2) {
-                    $fileSource = imagecreatefromjpeg($filename);
-                    //Rotates (if needed)
-                    if (isset($exifData['IFD0']['Orientation'])) {
-                        switch ($exifData['IFD0']['Orientation']) {
-                            case 1:
-                                $degree = 0;
-                                break;
-                            case 3:
-                                $degree = 180;
-                                break;
-                            case 6:
-                                $degree = 270;
-                                $newWidth = (int) round(($height * $newHeight) / $width);
-                                break;
-                            case 8:
-                                $degree = 90;
-                                $newWidth = (int) round(($height * $newHeight) / $width);
-                                break;
-                        }
-                        $fileSource = imagerotate($fileSource, $degree, 0);
-                    }
-                }
-                //PNG format
-                elseif ($fileData[2] == 3) {
-                    $fileSource = imagecreatefrompng($filename);
-                }
-
-                //Resizes
-                $newPicture = imagecreatetruecolor($newWidth, $newHeight);
-                if ($format == 'jpg') {
-                    $whiteBackground = imagecolorallocate($newPicture, 255, 255, 255);
-                    imagefill($newPicture, 0, 0, $whiteBackground);
-                }
-                if ($degree == 90 || $degree == 270) {
-                    imagecopyresampled($newPicture, $fileSource, 0, 0, 0, 0, $newWidth, $newHeight, $height, $width);
-                } else {
-                    imagecopyresampled($newPicture, $fileSource, 0, 0, 0, 0, $newWidth, $newHeight, $width, $height);
-                }
-
-                //Saves the picture - JPEG format
-                if ($format == 'jpg') {
-                    imagejpeg($newPicture, str_replace('jpeg', 'jpg', $filename), $compressionJpg);
-                }
-
-                //Destroy picture
-                imagedestroy($newPicture);
-
-                //Gets the FileSystem
-                $fs = new Filesystem();
-
-                //Saves the file in the right place
-                $folderPath = $this->getParameter('kernel.root_dir') . '/../web/images/' . $this->getParameter('c975_l_events.folderPictures');
-                $fs->mkdir($folderPath, 0770);
-                $file->move($folderPath, $finalFileName . '.jpg');
-            }
-        }
-    }
-
-    //Slugify function - https://github.com/cocur/slugify
-    public function slugify($text)
-    {
-        $slugify = new Slugify();
-        return $slugify->slugify($text);
     }
 }